A few more decoders in HighAttribute.
- decode_attr_synthetic
- decode_attr_runtime_invisible_annotations
- decode_attr_runtime_visible_annotations
- decode_attr_runtime_invisible_parameter_annotations
- decode_attr_runtime_visible_parameter_annotations

Also, treat StackMapTable annotation as unknown. In the ideal case, we
would verify that it corresponds to the bytecode, but we wouldn't put it
in the high-level representation anyway. During encoding, we should
reconstruct it thru symbolic execution. (We may not need this soon,
though.)

Now the whole JDK6 is decoded (in ~10s).