Various small fixes.
This version decodes JDK6 (without the malformed package-info.class)
and dacapo-9.12, on my computer.
- bumped upper version bound for JSR and RET to 1.6
- force the mapping in LineNumberTable to be injective, by forgetting
  (arbitrarily!) some entries
- be more lenient with class name checking

About class name checking: As far as I can tell there is no constraint
in the spec on how inner classes should be named. All that is said is that
what follows the last slash should BLA-BLA. So, the current check, that
splits the name on the first '$' and checks that both parts satisfy
BLA-BLA is still stronger than the spec requires. This should be solved
properly:
  https://github.com/rgrig/barista/issues/1