Attribute encoder stubs.
The structure is similar to decoding, with one interesting difference:
In the case of decoding the pattern is a string, which is why decoders
all go into a hash, which is why they must have the same type. For
encoding, however, the matching is a simple (polymorphic) variant pattern,
each branch calls a function, and those function are not forced to have
the same type. Still, we should probably refrain from making the types too
different.

I have also removed the explicit mutual recursion between [encode] and
[encode_attr_code]. I want the functions [encode_attr_*] to be sorted
alphabetically, but I don't want to declare all of them as being
mutually recursive. (I didn't see any argument more serious than code
aesthetics going one way or another.)