Rewrote [SymbExe.execute_method].
The current version does not use array or indexes within the instruction
list.  Interestingly, it now finishes quickly on the class over which it
took forever before, but it still takes forever on other examples.

I updated the code decoder to use HI.fresh_label().  This function makes
it easy for users to instrument bytecode. But this function does not work
correctly if we create labels also by other means.  While making this
change, I switched to int64 so we don't run out of labels too quickly.

The try_end field of the exception_table record used to point to the
label after the handled region.  Such [intervals) make sense with offsets,
but not so much with labels.  For example, there is no label pointing
immediately after the last instruction.  So I changed it to point to
the last instruction that is handled.  As a side-effect I changed the O(n)
conversion from offsets to labels into O(lg n).