Faster ExtendableArray.
This reduces the runtime of toplc (on jre and dacapo) by 40%. The old
linear search is now done faster. However, two things still need to be
done.

(1) ExtendableArray should be a functor of HashedType. (Right now I break
abstraction boundaries: I know that UTF8.equal is the same as =
because I looked at the implementation in camomile, but that's not in
the types.)

(2) ExtandableArray.index returns the smallest index, which is what
find did before.  The clients don't need this guarantee, AFAICT.  So
could this function be even faster?